requirements.txt
gunicorn
lxml==4.4.1
beautifulsoup4==4.8.1
bs4==0.0.1
pandas==0.23.0
requests==2.22.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.18.2
gspread==3.1.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
Werkzeug==0.16.0
